Update js-libp2p and ts-drp
allow float values for latency

+++ src/drpshadow/network/link.py
@@ -2,7 +2,7 @@ from drpshadow.network.zone import Zone
 
 
 class Link:
-    def __init__(self, src: Zone, dst: Zone, latency: int, packet_loss: float) -> None:
+    def __init__(self, src: Zone, dst: Zone, latency: float, packet_loss: float) -> None:
         self.src = src
         self.dst = dst
         self.latency = latency
